Refactored using BigDecimal instead of Double

+++ src/main/java/com/is/countryneighborstour/services/impl/ExchangeRateServiceImpl.java
@@ -5,6 +5,7 @@ import com.is.countryneighborstour.repositories.ExchangeRateRepository;
 import com.is.countryneighborstour.services.ExchangeRateService;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.time.LocalDate;
 import java.util.Map;
@@ -26,7 +27,7 @@ public class ExchangeRateServiceImpl implements ExchangeRateService {
     }
 
     @Override
-    public void saveExchangeRates(String baseCurrency, Map<String, Double> rates) {
+    public void saveExchangeRates(String baseCurrency, Map<String, BigDecimal> rates) {
         rates.forEach((targetCurrency, rate) -> {
             ExchangeRates exchangeRates = new ExchangeRates();
             exchangeRates.setBaseCurrency(baseCurrency);
